modified to find n primes but not efficient
This version does run for n primes up to 100000 and then gets a segmentation fault. I have realised that this is because the method to find the primes is not efficient memory wise and therefore not a suitable solution. I feel I have made it more complex than was anticipated.